+++ src/server/controllers/teachers/DeleteById.ts
@@ -1,10 +1,11 @@
-import { RequestHandler } from "express";
+import { Request, Response } from "express";
 import { TeachersProvider } from "../../database/providers/teachers/index.js";
-import { errorHandler } from "../../shared/errors/index.js";
+import { errorHandler, DataError } from "../../shared/errors/index.js";
 
-const deleteById: RequestHandler = async (req, res) => {
+const deleteById = async (req: Request, res: Response) => {
     try {
-        const id = Number(req.params.id);
+        const { id } = req.params;
+        if (!id) throw new DataError("invalid id");
         await TeachersProvider.deleteById(id);
         return res.status(200).json({ default: "teacher deleted" });
     } catch (err: unknown) {
